feat: sent envelopes from NDK on startup (#145)

+++ sentry-core/src/main/java/io/sentry/core/EnvelopeSender.java
@@ -1,5 +1,11 @@
 package io.sentry.core;
 
+import static io.sentry.core.ILogger.logIfNotNull;
+import static io.sentry.core.SentryLevel.ERROR;
+
+import io.sentry.core.hints.Cached;
+import io.sentry.core.hints.Retryable;
+import io.sentry.core.hints.SubmissionResult;
 import io.sentry.core.util.Objects;
 import java.io.ByteArrayInputStream;
 import java.io.File;
@@ -9,9 +15,12 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.nio.charset.Charset;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-public final class EnvelopeSender implements IEnvelopeSender {
+public final class EnvelopeSender extends DirectoryProcessor implements IEnvelopeSender {
 
   @SuppressWarnings("CharsetObjectCanBeUsed")
   private static final Charset UTF_8 = Charset.forName("UTF-8");
@@ -23,6 +32,7 @@ public final class EnvelopeSender implements IEnvelopeSender {
 
   public EnvelopeSender(
       IHub hub, IEnvelopeReader envelopeReader, ISerializer serializer, ILogger logger) {
+    super(logger);
     this.hub = Objects.requireNonNull(hub, "Hub is required.");
     this.envelopeReader = Objects.requireNonNull(envelopeReader, "Envelope reader is required.");
     this.serializer = Objects.requireNonNull(serializer, "Serializer is required.");
@@ -30,17 +40,20 @@ public final class EnvelopeSender implements IEnvelopeSender {
   }
 
   @Override
-  public void processEnvelopeFile(@NotNull String path) {
+  protected void processFile(@NotNull File file) {
     InputStream stream = null;
-    File file = null;
+    CachedEnvelopeHint hint =
+        new CachedEnvelopeHint(15000, logger); // TODO: Take timeout from options
     try {
-      file = new File(path);
       stream = new FileInputStream(file);
       SentryEnvelope envelope = envelopeReader.read(stream);
       if (envelope == null) {
-        logger.log(SentryLevel.ERROR, "Stream from path %s resulted in a null envelope.", path);
+        logger.log(
+            SentryLevel.ERROR,
+            "Stream from path %s resulted in a null envelope.",
+            file.getAbsolutePath());
       } else {
-        processEnvelope(envelope);
+        processEnvelope(envelope, hint);
       }
     } catch (Exception e) {
       logger.log(SentryLevel.ERROR, "Error processing envelope.", e);
@@ -50,8 +63,7 @@ public final class EnvelopeSender implements IEnvelopeSender {
       } catch (IOException ex) {
         logger.log(SentryLevel.ERROR, "Error closing envelope.", ex);
       }
-      if (file != null) {
-        // TODO: Handle error, at least ignore in memory
+      if (file != null && !hint.getRetry()) {
         try {
           file.delete();
         } catch (Exception e) {
@@ -61,11 +73,23 @@ public final class EnvelopeSender implements IEnvelopeSender {
     }
   }
 
-  private void processEnvelope(SentryEnvelope envelope) throws IOException {
+  @Override
+  protected boolean isRelevantFileName(String fileName) {
+    return true; // TODO: Use an extension to filter out relevant files
+  }
+
+  @Override
+  public void processEnvelopeFile(@NotNull String path) {
+    processFile(new File(path));
+  }
+
+  private void processEnvelope(@NotNull SentryEnvelope envelope, @NotNull CachedEnvelopeHint hint)
+      throws IOException {
     logger.log(SentryLevel.DEBUG, "Envelope for event Id: %s", envelope.getHeader().getEventId());
     int items = 0;
     for (SentryEnvelopeItem item : envelope.getItems()) {
       items++;
+
       if (item.getHeader() == null) {
         logger.log(SentryLevel.ERROR, "Item %d has no header", items);
         continue;
@@ -90,8 +114,16 @@ public final class EnvelopeSender implements IEnvelopeSender {
                   event.getEventId());
               continue;
             }
-            hub.captureEvent(event);
+            hub.captureEvent(event, hint);
             logger.log(SentryLevel.DEBUG, "Item %d is being captured.", items);
+            if (!hint.waitFlush()) {
+              logIfNotNull(
+                  logger,
+                  SentryLevel.WARNING,
+                  "Timed out waiting for event submission: %s",
+                  event.getEventId());
+              break;
+            }
           }
         }
       } else {
@@ -99,6 +131,63 @@ public final class EnvelopeSender implements IEnvelopeSender {
         logger.log(
             SentryLevel.WARNING, "Item %d of type: %s ignored.", items, item.getHeader().getType());
       }
+
+      if (!hint.succeeded) {
+        // Failed to send an item of the envelope: Stop attempting to send the rest (an attachment
+        // without the event that created it isn't useful)
+        logger.log(
+            SentryLevel.WARNING,
+            "Envelope for event Id: %s had a failed capture at item %d. No more items will be sent.",
+            envelope.getHeader().getEventId(),
+            items);
+        break;
+      }
+      hint.reset();
+    }
+  }
+
+  private static final class CachedEnvelopeHint implements Cached, Retryable, SubmissionResult {
+    boolean retry = false;
+    boolean succeeded = false;
+
+    private CountDownLatch latch;
+    private final long timeoutMills;
+    private final @Nullable ILogger logger;
+
+    CachedEnvelopeHint(final long timeoutMills, final @Nullable ILogger logger) {
+      this.timeoutMills = timeoutMills;
+      this.latch = new CountDownLatch(1);
+      this.logger = logger;
+    }
+
+    boolean waitFlush() {
+      try {
+        return latch.await(timeoutMills, TimeUnit.MILLISECONDS);
+      } catch (InterruptedException e) {
+        logIfNotNull(logger, ERROR, "Exception while awaiting on lock.", e);
+      }
+      return false;
+    }
+
+    public void reset() {
+      latch = new CountDownLatch(1);
+      succeeded = false;
+    }
+
+    @Override
+    public boolean getRetry() {
+      return retry;
+    }
+
+    @Override
+    public void setRetry(boolean retry) {
+      this.retry = retry;
+    }
+
+    @Override
+    public void setResult(boolean succeeded) {
+      this.succeeded = succeeded;
+      latch.countDown();
     }
   }
 }
